requirements.txt
streamlit>=1.32
pyvis>=0.3.2
duckdb>=0.10
pyyaml>=6.0
cerberus>=1.3
